backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["shraddhabhandarkar3 <bhandarkar.s@northeastern.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.10.15"
fastapi = "^0.100.0"  # Supports Pydantic v2
uvicorn = "^0.22.0"
openai = "^1.57.2"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
python-dotenv = "^1.0.0"
python-docx = "^0.8.11"
fpdf = "^1.7.2"
python-multipart = "^0.0.19"
cohere = "^5.13.3"
datasets = "2.19.1"
langchain-pinecone = "0.1.1"
langchain-openai = "0.1.9"
langchain = "0.2.5"
langchain-core = "0.2.9"
langgraph = "0.1.1"
semantic-router = "^0.0.72"  # Depends on Pydantic >=2.5.3
serpapi = "0.1.5"
google-search-results = "2.4.2"
ipython = "^8.29.0"
pygraphviz = "^1.14"
reportlab = "^4.2.5"
graphviz = "^0.20.3"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyodbc = "^5.2.0"
easyocr = "^1.7.2"
pypdf2 = "^3.0.1"
feedparser = "^6.0.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
